arm: tegra: tripndroid dvfs patches v1
tripndroid dvfs patches is a patchset for tegra 3 kernels with the latest
tegra dvfs system patched into it. this patchset prepares and updates the
kernel to a much more simple and easier dvfs structure. the basic goal is
to integrate this completely into tdf and restore original code if needed.
after patching has been done it will improve:

- performance
- battery lasting
- user experience

this initial patch cleans things up first.

if you have linear cpu age in the kernel then apply the previous patch
first, because this patch depends on a kernel without cpu linear age.

 Conflicts:
	arch/arm/mach-tegra/tegra3_dvfs.c