mmc: sdhci: Do not allow tuning procedure to be interrupted
wait_event_interruptible_timeout() will return early if the blocked
process receives a signal, causing the driver to abort the tuning
procedure and possibly leaving the controller in a bad state.  Since the
tuning command is expected to complete quickly (<50ms) and we've set a
timeout, use wait_event_timeout() instead.

BUG=chrome-os-partner:45463
TEST=Boot smaug and/or tune mmc devices repeatedly

Change-Id: Icceb104ee019bbf407990d86c6452bcf9aeecd65
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/304117
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Jared Dantis <jareddantis@gmail.com>