xfrm_user: validate XFRM_MSG_NEWAE incoming ESN size harder
Kees Cook has pointed out that xfrm_replay_state_esn_len()
is subject to wrapping issues. To ensure we are correctly
ensuring that the two ESN structures are the same size
compare both the overall size as reported by
xfrm_replay_state_esn_len() and the internal length are the same.

CVE-2017-7184

Signed-off-by: Jared Dantis <jareddantis@gmail.com>